Feature: Implement Dynamic Transaction Details Page
Added a dynamic page to render the details of each transaction. The committed code is a work in progress and currently displays only the transaction ID on the page. This serves as the initial stage of development for the transaction details page.

Please note that further enhancements and additional details will be incorporated in subsequent commits as the development progresses.