Add: Form and Prefilled Data in /edit page
Added a form to the /edit page and replaced the form values with
preexisting data when the id query string parameter is present. As of
this commit, the page allows users to view the form with prefilled data,
but no database call has been implemented to update the transaction yet.