Feature: Limit Fetching of Transactions
Implemented a fetch limit for transactions to optimize data retrieval and improve application performance. The limit is set to 5 for the home page and 12 for pagination, ensuring that transactions are fetched only when needed and reducing the load on the application. Additionally, a hard limit of 5 has been set for transactions displayed on the home page, providing a more focused and manageable view for users.

Note: Pagination functionality has not yet been implemented.