Feat: Implement Home Page and Access Restrictions
Added a new "home" page for the home-page of the app. This page is accessible only to new users (not signed-in). If a new user navigates away from this page, they will be automatically redirected back to the home page.

However, if a signed-in user navigates to this page, they will be redirected to the actual home page ("/") of the app. This access restriction ensures that new users are directed to the appropriate starting point of the application, while signed-in users are redirected to their personalized home page.