Fix: Resolve issue with webhook endpoint not hitting
Fixed an issue where the webhook endpoint (used when a new user is created) was not being triggered. Turns out that the API endpoint `/api/clerk` was not being hit for unknown reasons. To resolve this, the endpoint was renamed to `/webhook/clerk`, which successfully triggered the webhook and allowed new user profiles to be created for storing transaction data. Further investigation into this issue is planned to prevent similar occurrences in the future.