Feature: Implement Initial Balance Prompt Component
Added a new component that renders a warning and prompts the user to enter an initial balance if the starting balance is not set (value is currently hardcoded, will be fetched from the database in the future). However, if the initial balance is already set, the component does not render anything.