Feat: Implement edit functionality for initial balance
Added functionality to the Edit Initial Balance component so that clicking the edit button replaces the balance text with an input field that auto-focuses. When the input field loses focus, it reverts to displaying the balance text, discarding any changes made to the input value, effectively canceling the editing of the initial balance.

Note: Database is not connectied with this component, meaning editing
the initial component does nothing.